backend/pyproject.toml
[tool.poetry]
name = "Audio-dictionary"
version = "0.1.0"
description = ""
authors = ["Carla Prieto <carlapchavez@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.116.1"
uvicorn = "^0.35.0"
python-multipart = "^0.0.20"
websockets = "^15.0.1"
aiofiles = "^24.1.0"
httpx = "^0.28.1"
# openai = "^1.97.1"
python-dotenv = "^1.1.1"
ffmpeg-python = "^0.2.0"
imageio-ffmpeg = "^0.6.0"
numpy = "<2.0"
google-cloud-speech = "^2.33.0"
pydub = "^0.25.1"

# Compatible faster-whisper setup
# faster-whisper = { version = "^1.1.1", extras = [] }
# ctranslate2 = "^4.0"
# huggingface-hub = "0.16.4"
# tokenizers = "0.21.4"
# onnxruntime = "1.16.3"
# av = "11.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
